consolidate infer implemmentation in tests
